document-processing-classification/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# OCR and document processing
pytesseract>=0.3.0
pdf2image>=1.16.0
Pillow>=9.0.0
pypdf>=3.0.0
opencv-python>=4.5.0
python-magic>=0.4.0

# NLP and embeddings
transformers>=4.15.0
sentence-transformers>=2.2.0
torch>=1.9.0
spacy>=3.0.0
nltk>=3.6.0

# Clustering and visualization
# Removed umap-learn and hdbscan which require llvmlite
plotly>=5.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
joblib>=1.1.0
pyyaml>=6.0

# Development
jupyter>=1.0.0
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0